+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Include a compressed pickle file with all the archived data in this repo
 * `--load-new` option to scan for only new files
 * `--output-chrono` option to print all file types intermingled in chronological order
+* `--repair` option to repair a list of file IDs
 
 ### 1.6.1
 * site index at top that cross links all sites